+++ changelog.md
@@ -12,6 +12,7 @@
 * ManageAccountRule and ManageAccountRole op frames
 * AccountRole and AccountRule helpers and frames
 * Account frame
+* Create account op frame
 
 ### Deprecated
 
@@ -22,6 +23,7 @@
 
 * source account details, counterparty details
 * getSourceAccountDetails and getCounterpartyDetails for all op frames
+* manage account op frame
 
 ### Fixed